project/pyproject.toml
# there are different build systems which can be used
# but setuptools is old one, we can use that

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "package"
version = "0.0.1"
requires-python = ">=3.12"
readme = "README.md"

# there are ways to specify dependencies in a
# more specificway depending on OS
dependencies = [
  "pandas==2.2.3",
  "numpy==2.1.1",
  "python-json-logger==2.0.7",
  "torch==2.4.1",
  "torchvision==0.19.1",
  "PyYAML==6.0.2"
]

# to run a command in command line
# like in this case
# > package
# command will execute and run below function
[project.scripts]
packagepy = "package.__main__:main"
